Refactoring to have common code in SilxMaskImageWidget for PyMca5/PyMcaGui/pymca/SilxExternalImagesWindow.py and SilxStackRoiWindow.py
