+++ PyMca5/PyMcaGui/plotting/SilxMaskImageWidget.py
@@ -1,0 +1,240 @@
+# /*#########################################################################
+# Copyright (C) 2004-2017 V.A. Sole, European Synchrotron Radiation Facility
+#
+# This file is part of the PyMca X-ray Fluorescence Toolkit developed at
+# the ESRF by the Software group.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+#
+# ###########################################################################*/
+"""Base class for SilxStackRoiWindow and SilxExternalImagesWindow"""
+
+__authors__ = ["P. Knobel"]
+__license__ = "MIT"
+
+from PyMca5.PyMcaGui import PyMcaQt as qt
+if hasattr(qt, "QString"):
+    QString = qt.QString
+else:
+    QString = str
+from PyMca5.PyMcaGui.plotting.PyMca_Icons import IconDict
+
+from silx.gui.plot import PlotWidget
+from silx.gui.plot import PlotActions
+from silx.gui.plot import PlotToolButtons
+from silx.gui.plot import MaskToolsWidget
+
+
+class SilxMaskImageWidget(qt.QMainWindow):
+    """
+
+    """
+    sigMaskImageWidget = qt.pyqtSignal(object)
+
+    def __init__(self, parent=None):
+        qt.QMainWindow.__init__(self, parent=parent)
+        self.setWindowTitle("PyMca - Image Selection Tool")
+
+        centralWidget = qt.QWidget(self)
+        layout = qt.QVBoxLayout(centralWidget)
+        centralWidget.setLayout(layout)
+
+        # Plot
+        self.plot = PlotWidget(parent=centralWidget)
+        self.plot.setWindowFlags(qt.Qt.Widget)
+        layout.addWidget(self.plot)
+
+        # Mask Widget
+        self._maskToolsDockWidget = None
+
+        # ADD/REMOVE/REPLACE IMAGE buttons
+        # (methods and button connection to be implemented in subclass)
+        buttonBox = qt.QWidget(self)
+        buttonBoxLayout = qt.QHBoxLayout(buttonBox)
+        buttonBoxLayout.setContentsMargins(0, 0, 0, 0)
+        buttonBoxLayout.setSpacing(0)
+        self.addImageButton = qt.QPushButton(buttonBox)
+        icon = qt.QIcon(qt.QPixmap(IconDict["rgb16"]))
+        self.addImageButton.setIcon(icon)
+        self.addImageButton.setText("ADD IMAGE")
+        self.addImageButton.setToolTip("Add image to RGB correlator")
+        buttonBoxLayout.addWidget(self.addImageButton)
+
+        self.removeImageButton = qt.QPushButton(buttonBox)
+        self.removeImageButton.setIcon(icon)
+        self.removeImageButton.setText("REMOVE IMAGE")
+        self.removeImageButton.setToolTip("Remove image from RGB correlator")
+        buttonBoxLayout.addWidget(self.removeImageButton)
+
+        self.replaceImageButton = qt.QPushButton(buttonBox)
+        self.replaceImageButton.setIcon(icon)
+        self.replaceImageButton.setText("REPLACE IMAGE")
+        self.replaceImageButton.setToolTip(
+                "Replace all images in RGB correlator with this one")
+        buttonBoxLayout.addWidget(self.replaceImageButton)
+
+        layout.addWidget(buttonBox)
+
+        self.setCentralWidget(centralWidget)
+
+        # Init actions
+        self.group = qt.QActionGroup(self)
+        self.group.setExclusive(False)
+
+        self.resetZoomAction = self.group.addAction(
+                PlotActions.ResetZoomAction(plot=self.plot, parent=self))
+        self.addAction(self.resetZoomAction)
+
+        self.zoomInAction = PlotActions.ZoomInAction(plot=self.plot, parent=self)
+        self.addAction(self.zoomInAction)
+
+        self.zoomOutAction = PlotActions.ZoomOutAction(plot=self.plot, parent=self)
+        self.addAction(self.zoomOutAction)
+
+        self.yAxisInvertedButton = PlotToolButtons.YAxisOriginToolButton(
+            parent=self, plot=self.plot)
+        # self.yAxisInvertedButton.clicked.connect(self._hFlipIconSignal)
+
+        self.xAxisAutoScaleAction = self.group.addAction(
+            PlotActions.XAxisAutoScaleAction(plot=self.plot, parent=self))
+        self.addAction(self.xAxisAutoScaleAction)
+
+        self.yAxisAutoScaleAction = self.group.addAction(
+            PlotActions.YAxisAutoScaleAction(plot=self.plot, parent=self))
+        self.addAction(self.yAxisAutoScaleAction)
+
+        self.colormapAction = self.group.addAction(
+                PlotActions.ColormapAction(plot=self.plot, parent=self))
+        self.addAction(self.colormapAction)
+
+        self.keepDataAspectRatioButton = PlotToolButtons.AspectToolButton(
+            parent=self, plot=self.plot)
+
+        self.group.addAction(self.getMaskAction())
+
+        # Creating the toolbar also create actions for toolbuttons
+        self._toolbar = self._createToolBar(title='Plot', parent=None)
+        self.addToolBar(self._toolbar)
+
+    def sizeHint(self):
+        return qt.QSize(400, 400)
+
+    def _createToolBar(self, title, parent):
+        """Create a QToolBar with crop, rotate and flip operations
+
+        :param str title: The title of the QMenu
+        :param qt.QWidget parent: See :class:`QToolBar`
+        """
+        toolbar = qt.QToolBar(title, parent)
+
+        # Order widgets with actions
+        objects = self.group.actions()
+
+        # Add standard push buttons to list
+        index = objects.index(self.colormapAction)
+        objects.insert(index + 1, self.keepDataAspectRatioButton)
+        objects.insert(index + 2, self.yAxisInvertedButton)
+
+        for obj in objects:
+            if isinstance(obj, qt.QAction):
+                toolbar.addAction(obj)
+            else:
+                # keep reference to toolbutton's action for changing visibility
+                if obj is self.keepDataAspectRatioButton:
+                    self.keepDataAspectRatioAction = toolbar.addWidget(obj)
+                elif obj is self.yAxisInvertedButton:
+                    self.yAxisInvertedAction = toolbar.addWidget(obj)
+                else:
+                    raise RuntimeError()
+        return toolbar
+
+    def getMaskToolsDockWidget(self):
+        """DockWidget with image mask panel (lazy-loaded)."""
+        if self._maskToolsDockWidget is None:
+            self._maskToolsDockWidget = MaskToolsWidget.MaskToolsDockWidget(
+                plot=self.plot, name='Mask')
+            self._maskToolsDockWidget.hide()
+            self.addDockWidget(qt.Qt.BottomDockWidgetArea,
+                               self._maskToolsDockWidget)
+            self._maskToolsDockWidget.setFloating(True)
+            self._maskToolsDockWidget.sigMaskChanged.connect(
+                    self._emitMaskImageWidgetSignal)
+        return self._maskToolsDockWidget
+
+    def getMaskAction(self):
+        """QAction toggling image mask dock widget
+
+        :rtype: QAction
+        """
+        return self.getMaskToolsDockWidget().toggleViewAction()
+
+    def _emitMaskImageWidgetSignal(self):
+        self.sigMaskImageWidget.emit(
+            {"event": "selectionMaskChanged",
+             "current": self.getSelectionMask(),
+             "id": id(self)})
+
+    def setSelectionMask(self, mask, copy=True):
+        """Set the mask to a new array.
+
+        :param numpy.ndarray mask: The array to use for the mask.
+                    Mask type: array of uint8 of dimension 2,
+                    Array of other types are converted.
+        :param bool copy: True (the default) to copy the array,
+                          False to use it as is if possible.
+        :return: None if failed, shape of mask as 2-tuple if successful.
+                 The mask can be cropped or padded to fit active image,
+                 the returned shape is that of the active image.
+        """
+        # don't emit signal for programmatic mask change,
+        # only for interactive mask drawing
+        # (avoid infinite loop)
+        self.getMaskToolsDockWidget().sigMaskChanged.disconnect(
+                    self._emitMaskImageWidgetSignal)
+        ret = self.getMaskToolsDockWidget().setSelectionMask(mask,
+                                                             copy=copy)
+        self.getMaskToolsDockWidget().sigMaskChanged.connect(
+                    self._emitMaskImageWidgetSignal)
+        return ret
+
+    def getSelectionMask(self, copy=True):
+        """Get the current mask as a 2D array.
+
+        :param bool copy: True (default) to get a copy of the mask.
+                          If False, the returned array MUST not be modified.
+        :return: The array of the mask with dimension of the 'active' image.
+                 If there is no active image, an empty array is returned.
+        :rtype: 2D numpy.ndarray of uint8
+        """
+        return self.getMaskToolsDockWidget().getSelectionMask(copy=copy)
+    #
+    # def _hFlipIconSignal(self):
+    #     # inform the other widgets
+    #     self.sigMaskImageWidget.emit(
+    #         {'event': "hFlipSignal",
+    #          'current': self.plot.isYAxisInverted(),
+    #          'id': id(self)})
+
+
+if __name__ == "__main__":
+    app = qt.QApplication([])
+    w = SilxMaskImageWidget()
+    w.show()
+    w.plot.addImage([[0, 1, 2], [2, 1, -1]])
+    app.exec_()
